Rework project structure
