+++ pkg/tools/api/resourcepool/host_info.go
