+++ pkg/tools/api/resourcepool/resource_pool.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2020 Doc.ai and/or its affiliates.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package resourcepool provides types for the PCI resource pool api
+package resourcepool
+
+import "github.com/networkservicemesh/sdk-sriov/pkg/tools/api/pcifunction"
+
+// HostInfoProvider provides host info
+type HostInfoProvider interface {
+	GetHostInfo() *HostInfo
+}
+
+// ResourcePool provides methods to select and free PCI functions
+type ResourcePool interface {
+	Select(pfPciAddr string, igid uint, driverType DriverType) (pcifunction.PCIFunction, error)
+	SelectAny(pfPciAddr string, driverType DriverType) (pcifunction.PCIFunction, error)
+	Free(vfPciAddr string)
+}
