+++ test/tools/resourcepool/host_info.yml
